feat(router): 引入 Vue Router，为各视图添加独立路由
- 新增 vue-router@4 依赖
- 新增 src/composables/useAuth.js：模块单例共享登录状态
- 新增 src/router/index.js：history 模式路由 + 导航守卫
  - /auth → 登录/注册页
  - /app/:view?/:subview? → 工作台（需登录）
  - /app/workspace/review 独立为题目核对页路由
- App.vue 重构为薄壳 RouterView，负责主题初始化
- WorkspaceView.vue：从 App.vue 迁移工作台内容
  - currentView 改为 writable computed，自动同步 URL
  - handleLogout 后跳转 /auth
- AuthView.vue：登录/注册成功后 router.push('/app/workspace')
- vite.config.js：添加开发模式中间件，重写 /auth 和 /app/* 到 app.html
- web_app.py：新增 /auth 路由及 /app/<path> 通配，均返回 app.html
