+++ frontend/package.json
@@ -13,7 +13,8 @@
   "dependencies": {
     "@headlessui/vue": "^1.7.23",
     "dompurify": "^3.3.2",
-    "vue": "^3.5.24"
+    "vue": "^3.5.24",
+    "vue-router": "^4.6.4"
   },
   "devDependencies": {
     "@vitejs/plugin-vue": "^6.0.1",
